requirements.txt
asgiref==3.7.2
Django==3.2.4
django-ckeditor==6.7.0
django-js-asset==2.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==10.0.0
psycopg2-binary==2.9.3
PyJWT==2.8.0
python-decouple==3.4
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
